Moved hader and cell to other file.
Issue #68.
