+++ GiniVision/Classes/CaptureSuggestionsCollectionView.swift
@@ -78,88 +78,3 @@ final class CaptureSuggestionsCollectionView: UICollectionView {
     
 }
 
-final class CaptureSuggestionsCollectionCell: UICollectionViewCell {
-    
-    var suggestionImage: UIImageView = {
-        let suggestionImage = UIImageView()
-        suggestionImage.translatesAutoresizingMaskIntoConstraints = false
-        suggestionImage.contentMode = .scaleAspectFit
-        return suggestionImage
-    }()
-    var suggestionText: UILabel = {
-        let suggestionText = UILabel()
-        suggestionText.translatesAutoresizingMaskIntoConstraints = false
-        suggestionText.numberOfLines = 0
-        suggestionText.adjustsFontSizeToFitWidth = true
-        suggestionText.font = UIFont.systemFont(ofSize: 14)
-        suggestionText.minimumScaleFactor = 10 / 14
-        return suggestionText
-    }()
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubview(suggestionImage)
-        addSubview(suggestionText)
-        addConstraints()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(frame:) should be used instead")
-    }
-    
-    private func addConstraints() {
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .top, relatedBy: .equal, toItem: self, attribute: .top, multiplier: 1.0, constant: 0, priority: 999)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .bottom, relatedBy: .equal, toItem: self, attribute: .bottom, multiplier: 1.0, constant: 0, priority: 999)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .leading, relatedBy: .equal, toItem: self, attribute: .leading, multiplier: 1.0, constant: 20)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .trailing, relatedBy: .equal, toItem: suggestionText, attribute: .leading, multiplier: 1.0, constant: -20)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .width, relatedBy: .equal, toItem: nil, attribute: .notAnAttribute, multiplier: 1.0, constant: 85)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .height, relatedBy: .lessThanOrEqual, toItem: nil, attribute: .notAnAttribute, multiplier: 1.0, constant: 85)
-        ConstraintUtils.addActiveConstraint(item: suggestionImage, attribute: .centerY, relatedBy: .equal, toItem: self, attribute: .centerY, multiplier: 1.0, constant: 0)
-        
-        ConstraintUtils.addActiveConstraint(item: suggestionText, attribute: .top, relatedBy: .equal, toItem: self, attribute: .top, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: suggestionText, attribute: .bottom, relatedBy: .equal, toItem: self, attribute: .bottom, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: suggestionText, attribute: .trailing, relatedBy: .equal, toItem: self, attribute: .trailing, multiplier: 1.0, constant: -20, priority: 999)
-    }
-}
-
-final class CaptureSuggestionsCollectionHeader: UICollectionReusableView {
-    
-    var headerTitle: UILabel = {
-        let headerTitle = UILabel()
-        headerTitle.translatesAutoresizingMaskIntoConstraints = false
-        headerTitle.numberOfLines = 0
-        headerTitle.font = UIFont.boldSystemFont(ofSize: 14)
-        return headerTitle
-    }()
-    var bottomLine: UIView = {
-        let bottomLine = UIView()
-        bottomLine.translatesAutoresizingMaskIntoConstraints = false
-        bottomLine.backgroundColor = .lightGray
-        return bottomLine
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        backgroundColor = .white
-        addSubview(headerTitle)
-        addSubview(bottomLine)
-        addConstraints()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(frame:) should be used instead")
-    }
-    
-    private func addConstraints() {
-        ConstraintUtils.addActiveConstraint(item: headerTitle, attribute: .top, relatedBy: .equal, toItem: self, attribute: .top, multiplier: 1.0, constant: 20)
-        ConstraintUtils.addActiveConstraint(item: headerTitle, attribute: .leading, relatedBy: .equal, toItem: self, attribute: .leading, multiplier: 1.0, constant: 20)
-        ConstraintUtils.addActiveConstraint(item: headerTitle, attribute: .trailing, relatedBy: .equal, toItem: self, attribute: .trailing, multiplier: 1.0, constant: -20)
-        ConstraintUtils.addActiveConstraint(item: headerTitle, attribute: .bottom, relatedBy: .equal, toItem: bottomLine, attribute: .top, multiplier: 1.0, constant: -20)
-        
-        ConstraintUtils.addActiveConstraint(item: bottomLine, attribute: .bottom, relatedBy: .equal, toItem: self, attribute: .bottom, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: bottomLine, attribute: .leading, relatedBy: .equal, toItem: self, attribute: .leading, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: bottomLine, attribute: .trailing, relatedBy: .equal, toItem: self, attribute: .trailing, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: bottomLine, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .notAnAttribute, multiplier: 1.0, constant: 0.0)
-        
-    }
-}
-
